Move WithSignalManagerMixin and WithSignalQuerysetMixin to common/models/mixins
